+++ base/deprecated.jl
@@ -1354,6 +1354,8 @@ export conv, conv2, deconv, filt, filt!, xcorr
 @deprecate_moved watch_file "FileWatching" true true
 @deprecate_moved FileMonitor "FileWatching" true true
 
+@deprecate_moved crc32c "CRC32C" true true
+
 # PR #21709
 @deprecate cov(x::AbstractVector, corrected::Bool) cov(x, corrected=corrected)
 @deprecate cov(x::AbstractMatrix, vardim::Int, corrected::Bool) cov(x, vardim, corrected=corrected)
