+++ base/util.jl
@@ -713,44 +713,14 @@ if Sys.iswindows()
 
 end
 
-"""
-    crc32c(data, crc::UInt32=0x00000000)
-
-Compute the CRC-32c checksum of the given `data`, which can be
-an `Array{UInt8}`, a contiguous subarray thereof, or a `String`.  Optionally, you can pass
-a starting `crc` integer to be mixed in with the checksum.  The `crc` parameter
-can be used to compute a checksum on data divided into chunks: performing
-`crc32c(data2, crc32c(data1))` is equivalent to the checksum of `[data1; data2]`.
-(Technically, a little-endian checksum is computed.)
-
-There is also a method `crc32c(io, nb, crc)` to checksum `nb` bytes from
-a stream `io`, or `crc32c(io, crc)` to checksum all the remaining bytes.
-Hence you can do [`open(crc32c, filename)`](@ref) to checksum an entire file,
-or `crc32c(seekstart(buf))` to checksum an [`IOBuffer`](@ref) without
-calling [`take!`](@ref).
-
-For a `String`, note that the result is specific to the UTF-8 encoding
-(a different checksum would be obtained from a different Unicode encoding).
-To checksum an `a::Array` of some other bitstype, you can do `crc32c(reinterpret(UInt8,a))`,
-but note that the result may be endian-dependent.
-"""
-function crc32c end
-
 unsafe_crc32c(a, n, crc) = ccall(:jl_crc32c, UInt32, (UInt32, Ptr{UInt8}, Csize_t), crc, a, n)
 
-crc32c(a::Union{Array{UInt8},FastContiguousSubArray{UInt8,N,<:Array{UInt8}} where N}, crc::UInt32=0x00000000) =
+_crc32c(a::Union{Array{UInt8},FastContiguousSubArray{UInt8,N,<:Array{UInt8}} where N}, crc::UInt32=0x00000000) =
     unsafe_crc32c(a, length(a) % Csize_t, crc)
 
-crc32c(s::String, crc::UInt32=0x00000000) = unsafe_crc32c(s, sizeof(s) % Csize_t, crc)
+_crc32c(s::String, crc::UInt32=0x00000000) = unsafe_crc32c(s, sizeof(s) % Csize_t, crc)
 
-"""
-    crc32c(io::IO, [nb::Integer,] crc::UInt32=0x00000000)
-
-Read up to `nb` bytes from `io` and return the CRC-32c checksum, optionally
-mixed with a starting `crc` integer.  If `nb` is not supplied, then
-`io` will be read until the end of the stream.
-"""
-function crc32c(io::IO, nb::Integer, crc::UInt32=0x00000000)
+function _crc32c(io::IO, nb::Integer, crc::UInt32=0x00000000)
     nb < 0 && throw(ArgumentError("number of bytes to checksum must be ≥ 0"))
     # use block size 24576=8192*3, since that is the threshold for
     # 3-way parallel SIMD code in the underlying jl_crc32c C function.
@@ -762,8 +732,8 @@ function crc32c(io::IO, nb::Integer, crc::UInt32=0x00000000)
     end
     return unsafe_crc32c(buf, readbytes!(io, buf, min(nb, length(buf))), crc)
 end
-crc32c(io::IO, crc::UInt32=0x00000000) = crc32c(io, typemax(Int64), crc)
-crc32c(io::IOStream, crc::UInt32=0x00000000) = crc32c(io, filesize(io)-position(io), crc)
+_crc32c(io::IO, crc::UInt32=0x00000000) = _crc32c(io, typemax(Int64), crc)
+_crc32c(io::IOStream, crc::UInt32=0x00000000) = _crc32c(io, filesize(io)-position(io), crc)
 
 
 """
