+++ doc/make.jl
@@ -22,20 +22,22 @@ cp_q(src, dest) = isfile(dest) || cp(src, dest)
 # make links for stdlib package docs
 if Sys.iswindows()
     cp_q("../stdlib/DelimitedFiles/docs/src/index.md", "src/stdlib/delimitedfiles.md")
-    cp_q("../stdlib/Test/docs/src/index.md", "src/stdlib/test.md")
-    cp_q("../stdlib/Mmap/docs/src/index.md", "src/stdlib/mmap.md")
-    cp_q("../stdlib/SharedArrays/docs/src/index.md", "src/stdlib/sharedarrays.md")
-    cp_q("../stdlib/Profile/docs/src/index.md", "src/stdlib/profile.md")
-    cp_q("../stdlib/Base64/docs/src/index.md", "src/stdlib/base64.md")
-    cp_q("../stdlib/FileWatching/docs/src/index.md", "src/stdlib/filewatching.md")
+    cp_q("../stdlib/Test/docs/src/index.md",           "src/stdlib/test.md")
+    cp_q("../stdlib/Mmap/docs/src/index.md",           "src/stdlib/mmap.md")
+    cp_q("../stdlib/SharedArrays/docs/src/index.md",   "src/stdlib/sharedarrays.md")
+    cp_q("../stdlib/Profile/docs/src/index.md",        "src/stdlib/profile.md")
+    cp_q("../stdlib/Base64/docs/src/index.md",         "src/stdlib/base64.md")
+    cp_q("../stdlib/FileWatching/docs/src/index.md",   "src/stdlib/filewatching.md")
+    cp_q("../stdlib/CRC32c/docs/src/index.md",         "src/stdlib/crc32c.md")
 else
     symlink_q("../../../stdlib/DelimitedFiles/docs/src/index.md", "src/stdlib/delimitedfiles.md")
-    symlink_q("../../../stdlib/Test/docs/src/index.md", "src/stdlib/test.md")
-    symlink_q("../../../stdlib/Mmap/docs/src/index.md", "src/stdlib/mmap.md")
-    symlink_q("../../../stdlib/SharedArrays/docs/src/index.md", "src/stdlib/sharedarrays.md")
-    symlink_q("../../../stdlib/Profile/docs/src/index.md", "src/stdlib/profile.md")
-    symlink_q("../../../stdlib/Base64/docs/src/index.md", "src/stdlib/base64.md")
-    symlink_q("../../../stdlib/FileWatching/docs/src/index.md", "src/stdlib/filewatching.md")
+    symlink_q("../../../stdlib/Test/docs/src/index.md",           "src/stdlib/test.md")
+    symlink_q("../../../stdlib/Mmap/docs/src/index.md",           "src/stdlib/mmap.md")
+    symlink_q("../../../stdlib/SharedArrays/docs/src/index.md",   "src/stdlib/sharedarrays.md")
+    symlink_q("../../../stdlib/Profile/docs/src/index.md",        "src/stdlib/profile.md")
+    symlink_q("../../../stdlib/Base64/docs/src/index.md",         "src/stdlib/base64.md")
+    symlink_q("../../../stdlib/FileWatching/docs/src/index.md",   "src/stdlib/filewatching.md")
+    symlink_q("../../../stdlib/CRC32c/docs/src/index.md",         "src/stdlib/crc32c.md")
 end
 
 const PAGES = [
@@ -109,6 +111,7 @@ const PAGES = [
         "stdlib/mmap.md",
         "stdlib/sharedarrays.md",
         "stdlib/filewatching.md",
+        "stdlib/crc32c.md",
     ],
     "Developer Documentation" => [
         "devdocs/reflection.md",
@@ -143,7 +146,7 @@ const PAGES = [
     ],
 ]
 
-using DelimitedFiles, Test, Mmap, SharedArrays, Profile, Base64, FileWatching
+using DelimitedFiles, Test, Mmap, SharedArrays, Profile, Base64, FileWatching, CRC32c
 
 makedocs(
     build     = joinpath(pwd(), "_build/html/en"),
