+++ doc/src/stdlib/.gitignore
@@ -3,3 +3,7 @@ test.md
 mmap.md
 sharedarrays.md
 profile.md
+base64.md
+filewatching.md
+crc32c.md
+
