+++ doc/src/stdlib/arrays.md
@@ -134,7 +134,6 @@ Base.cumprod!
 Base.cumsum
 Base.cumsum!
 Base.cumsum_kbn
-Base.crc32c
 Base.LinAlg.diff
 Base.repeat(::AbstractArray)
 Base.rot180
