+++ stdlib/CRC32c/docs/src/index.md
@@ -1,0 +1,6 @@
+# CRC32c
+
+```@docs
+CRC32c.crc32c
+CRC32c.crc32c(::IO, ::Integer, ::UInt32)
+```
